TodayView 이모지 중복 선택 제외
