+++ client/MEWE/MEWE/Utility/Emojis.swift
@@ -1,0 +1,12 @@
+//
+//  Emojis.swift
+//  MEWE
+//
+//  Created by 1v1 on 2021/02/27.
+//
+
+import Foundation
+
+let emojiStrings = ["emoji_anger","emoji_concern", "emoji_ease", "emoji_flustration", "emoji_happiness", "emoji_heart", "emoji_joy", "emoji_proud", "emoji_sadness", "emoji_tiredness"]
+
+let emojiDetails = ["분노", "걱정", "편안", "당황", "행복", "설렘", "즐거움", "뿌듯", "슬픔", "피곤"]
